OrgPosition Final

+++ NHCM.Application/Organogram/Models/SearchedOrganogramControll.cs
@@ -8,7 +8,7 @@ namespace NHCM.Application.Organogram.Models
     {
         public int Id { get; set; }
         public int OrganizationId { get; set; }
-        public short StatusId { get; set; }
+        public int StatusId { get; set; }
         public DateTime ModifiedOn { get; set; }
         public string ModifiedBy { get; set; }
         public DateTime CreatedOn { get; set; }
